create separate component and added where applies
