+++ app/routes/_auth+/login.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 
-import { useState } from "react";
 import type { ActionArgs, LoaderArgs, V2_MetaFunction } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import {
@@ -13,7 +12,7 @@ import { parseFormAny, useZorm } from "react-zorm";
 import { z } from "zod";
 
 import Input from "~/components/forms/input";
-import { EyeIcon, EyeOffIcon } from "~/components/icons";
+import PasswordInput from "~/components/forms/password-input";
 import { Button } from "~/components/shared/button";
 
 import {
@@ -99,12 +98,6 @@ export default function IndexLoginForm() {
   const navigation = useNavigation();
   const disabled = isFormProcessing(navigation.state);
 
-  const [showPassword, setShowPassword] = useState(false);
-
-  const togglePasswordVisibility = () => {
-    setShowPassword(!showPassword);
-  };
-
   return (
     <div className="w-full max-w-md">
       <Form ref={zo.ref} method="post" replace className="flex flex-col gap-5">
@@ -123,30 +116,16 @@ export default function IndexLoginForm() {
             error={zo.errors.email()?.message}
           />
         </div>
-
-        <div className="relative">
-          <Input
-            label="Password"
-            placeholder="**********"
-            data-test-id="password"
-            name={zo.fields.password()}
-            type={showPassword ? "text" : "password"}
-            autoComplete="new-password"
-            disabled={disabled}
-            inputClassName="w-full"
-            error={zo.errors.password()?.message || data?.errors?.email}
-          />
-          <span
-            className="absolute bottom-2  right-2 flex h-6 w-6 cursor-pointer flex-col items-end justify-center"
-            onClick={togglePasswordVisibility}
-          >
-            {showPassword ? (
-              <EyeOffIcon className="h-full w-full" />
-            ) : (
-              <EyeIcon className="h-full w-full" />
-            )}
-          </span>
-        </div>
+        <PasswordInput
+          label="Password"
+          placeholder="**********"
+          data-test-id="password"
+          name={zo.fields.password()}
+          autoComplete="new-password"
+          disabled={disabled}
+          inputClassName="w-full"
+          error={zo.errors.password()?.message || data?.errors?.email}
+        />
 
         <input type="hidden" name={zo.fields.redirectTo()} value={redirectTo} />
         <Button
